# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DOSTAC website + admin API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /public/products:
    get:
      operationId: listPublicProducts
      tags: [public-products]
      summary: List published products in a given language
      parameters:
        - { name: lang, in: query, required: true, schema: { $ref: "#/components/schemas/Lang" } }
        - { name: category, in: query, required: false, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/PublicProduct" }
  /public/products/{slug}:
    get:
      operationId: getPublicProduct
      tags: [public-products]
      parameters:
        - { name: slug, in: path, required: true, schema: { type: string } }
        - { name: lang, in: query, required: true, schema: { $ref: "#/components/schemas/Lang" } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PublicProduct" }
        "404": { $ref: "#/components/responses/NotFound" }
  /public/notices:
    get:
      operationId: listPublicNotices
      tags: [public-notices]
      parameters:
        - { name: lang, in: query, required: true, schema: { $ref: "#/components/schemas/Lang" } }
        - { name: category, in: query, required: false, schema: { type: string } }
        - { name: search, in: query, required: false, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/PublicNotice" }
  /public/notices/{slug}:
    get:
      operationId: getPublicNotice
      tags: [public-notices]
      parameters:
        - { name: slug, in: path, required: true, schema: { type: string } }
        - { name: lang, in: query, required: true, schema: { $ref: "#/components/schemas/Lang" } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PublicNotice" }
        "404": { $ref: "#/components/responses/NotFound" }
  /public/contact-inquiries:
    post:
      operationId: createContactInquiry
      tags: [public-contact]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ContactInquiryInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }
        "400": { $ref: "#/components/responses/BadRequest" }

  # ───────────────────── ADMIN AUTH ─────────────────────
...
  /admin/uploads/sign:
    post:
      operationId: adminSignUpload
      tags: [admin-uploads]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UploadSignInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UploadSignResult" }
components:
  responses:
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
...
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
  schemas:
...
    Ok:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean }
    ApiError:
      type: object
      required: [error]
      properties:
        error: { type: string }
    Lang:
      type: string
      enum: [ko, en, ja, zh, vi]
...
    PublicProduct:
      type: object
      required: [id, slug, category, sortOrder, name, headline, body]
      properties:
        id: { type: integer }
        slug: { type: string }
        category: { type: string }
        sortOrder: { type: integer }
        imageUrl: { type: ["string", "null"] }
        name: { type: string }
        headline: { type: string }
        body: { type: string }
    PublicNotice:
      type: object
      required: [id, slug, category, region, publishedAt, title, excerpt, body]
      properties:
        id: { type: integer }
        slug: { type: string }
        category: { type: string }
        region: { type: string }
        publishedAt: { type: string, format: date-time }
        thumbnailUrl: { type: ["string", "null"] }
        title: { type: string }
        excerpt: { type: string }
        body: { type: string }
...
    ContactInquiryInput:
      type: object
      required: [company, name, email, message]
      properties:
        company: { type: string }
        name: { type: string }
        email: { type: string }
        phone: { type: string }
        country: { type: string }
        projectType: { type: string }
        productInterest:
          type: array
          items: { type: string }
        monthlyVolume: { type: string }
        message: { type: string }
...
    UploadSignInput:
      type: object
      required: [filename, contentType]
      properties:
        filename: { type: string }
        contentType: { type: string }
    UploadSignResult:
      type: object
      required: [uploadUrl, publicUrl, key]
      properties:
        uploadUrl: { type: string }
        publicUrl: { type: string }
        key: { type: string }