# lib/api-zod/src/generated/types/publicProduct.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DOSTAC website + admin API
 * OpenAPI spec version: 0.1.0
 */

export interface PublicProduct {
  id: number;
  slug: string;
  category: string;
  sortOrder: number;
  /** @nullable */
  imageUrl?: string | null;
  name: string;
  headline: string;
  body: string;
}